pyproject.toml
[project]
name = "K-Ops"
version = "0.1.0"
description = "Starter agent-first Markdown knowledge base."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "beautifulsoup4>=4.12.0",
  "pypdf>=5.0.0",
  "pyyaml>=6.0.0",
  "requests>=2.32.0",
  "trafilatura>=1.9.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["scripts"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
package = false
